Support single region and add non-abstract editors
