--- Bonsai.Vision.Design/IplImageEllipseRoiEditor.cs
@@ -1,94 +1,0 @@
-﻿using System;
-using System.Linq;
-using System.Drawing.Design;
-using System.ComponentModel;
-using System.Windows.Forms.Design;
-using Bonsai.Design;
-using OpenCV.Net;
-using System.Reactive.Linq;
-using System.Windows.Forms;
-
-namespace Bonsai.Vision.Design
-{
-    /// <summary>
-    /// Provides an abstract base class for user interface editors that allow
-    /// visually editing a collection of elliptical regions on top of the active
-    /// image source.
-    /// </summary>
-    public abstract class IplImageEllipseRoiEditor : DataSourceTypeEditor
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="IplImageEllipseRoiEditor"/> class
-        /// using the specified image data source.
-        /// </summary>
-        /// <param name="source">
-        /// Specifies the source of image notifications to the property editor.
-        /// </param>
-        protected IplImageEllipseRoiEditor(DataSource source)
-            : base(source, typeof(IplImage))
-        {
-        }
-
-        /// <summary>
-        /// Gets the sequence of images arriving to or from the operator.
-        /// </summary>
-        /// <param name="source">
-        /// An observable sequence that multicasts notifications from all the active
-        /// subscriptions to the workflow operator.
-        /// </param>
-        protected virtual IObservable<IplImage> GetImageSource(IObservable<IObservable<object>> source)
-        {
-            return source.Merge().Select(image => image as IplImage);
-        }
-
-        /// <inheritdoc/>
-        public override UITypeEditorEditStyle GetEditStyle(ITypeDescriptorContext context)
-        {
-            return UITypeEditorEditStyle.Modal;
-        }
-
-        /// <inheritdoc/>
-        public override object EditValue(ITypeDescriptorContext context, IServiceProvider provider, object value)
-        {
-            var editorService = (IWindowsFormsEditorService)provider.GetService(typeof(IWindowsFormsEditorService));
-            if (context != null && editorService != null)
-            {
-                using (var visualizerDialog = new TypeVisualizerDialog())
-                {
-                    var imageControl = new ImageEllipseRoiPicker();
-                    var propertyDescriptor = context.PropertyDescriptor;
-                    var regions = (RotatedRect[])value;
-
-                    imageControl.Dock = DockStyle.Fill;
-                    visualizerDialog.Text = propertyDescriptor.Name;
-                    if (regions != null)
-                    {
-                        foreach (var region in regions) imageControl.Regions.Add(region);
-                    }
-
-                    imageControl.RegionsChanged += (sender, e) => propertyDescriptor.SetValue(context.Instance, imageControl.Regions.ToArray());
-                    visualizerDialog.AddControl(imageControl);
-                    imageControl.Canvas.MouseDoubleClick += (sender, e) =>
-                    {
-                        if (e.Button == MouseButtons.Left && imageControl.Image != null && !imageControl.SelectedRegion.HasValue)
-                        {
-                            visualizerDialog.ClientSize = new System.Drawing.Size(
-                                imageControl.Image.Width,
-                                imageControl.Image.Height);
-                        }
-                    };
-
-                    IDisposable subscription = null;
-                    var source = GetDataSource(context, provider);
-                    var imageSource = GetImageSource(source.Output);
-                    imageControl.Load += delegate { subscription = imageSource.Subscribe(image => imageControl.Image = image); };
-                    imageControl.HandleDestroyed += delegate { subscription.Dispose(); };
-                    editorService.ShowDialog(visualizerDialog);
-                    return imageControl.Regions.ToArray();
-                }
-            }
-
-            return base.EditValue(context, provider, value);
-        }
-    }
-}
